requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
flask-login==0.5.0
python-dotenv==0.19.0
werkzeug==2.0.1
flask-wtf==0.15.1
flask-bcrypt==0.7.1
flask-admin==1.6.0
email-validator==1.1.3
openpyxl==3.1.5
pytz==2024.1
gunicorn==20.1.0
